Code cleanups
git-svn-id: https://svn.atnf.csiro.au/difx@6491 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ libraries/mark5access/trunk/examples/m5fb.c
@@ -556,7 +556,7 @@ int spec(const char *filename, const char *formatname, int nchan, int nint, cons
 				}
 			}
 			ssize_t nwr = fwrite(tmp,sizeof(char),nchan*nif,out);
-			printf("%d/%d samples: \r", nwr, count++);
+			printf("%ld/%d samples: \r", (long)nwr, count++);
 			//fwrite(tmp,sizeof(char),nchan*4,out);
 		}
 		else
